refactor(Request): Remove `getScriptUrl()` method for `ServerRequestAdapter` class and related tests; introduce `getScriptName()` for improved script URL handling in PSR-7 and Yii2 environments.
